piggyback_insider_trading/requirements.txt
sqlalchemy>=2
pydantic>=2
python-dotenv
requests
lxml
apscheduler
psycopg[binary]
